% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar1_cov}
\alias{ar1_cov}
\title{Covariance matrix for a stationary Gaussian AR(1) process.}
\usage{
ar1_cov(times, rho, sigma)
}
\arguments{
\item{times}{An vector of positive integers, preferably ordered.}

\item{rho}{A real number strictly less than 1 in absolute value.}

\item{sigma}{A positive real number.}
}
\value{
A square matrix with \code{length(times)} rows if \code{times} has
  more than 1 element. Otherwise, a matrix with \code{length(times[1])}
  elements.
}
\description{
Creates the covariance matrix of an AR(1) process with parameters \code{rho}
and \code{sigma}, observed at the time points in the vector \code{times}.
The process is assumed to be in stationarity and to have Gaussian errors.
If times is a vector of length 1, and its value is a positive integer
\code{n}, then the covariance matrix for \code{n} consecutive time points
is created instead.
}
\examples{
times <- c(1, 4:5, 7)
rho <- 0.5
sigma <- 1
ar1_cov(times, rho, sigma)
}
