% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar1_cov_chol_irregular}
\alias{ar1_cov_chol_irregular}
\title{Upper triangular Cholesky factorization of covariance matrix of a vector x
observed at times times.}
\usage{
ar1_cov_chol_irregular(times, rho, sigma)
}
\description{
Upper triangular Cholesky factorization of covariance matrix of a vector x
observed at times times.
}
