% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar1_cross_cov}
\alias{ar1_cross_cov}
\title{Partial covariance matrix of vectors x and y observed at  (irregularly
spaced) times times1 and times2, respectively.}
\usage{
ar1_cross_cov(times1, times2, rho, sigma)
}
\arguments{
\item{times1}{An vector of positive integers, preferably ordered.}

\item{times2}{An vector of positive integers, preferably ordered.}

\item{rho}{A real number strictly less than 1 in absolute value.}

\item{sigma}{A positive real number.}
}
\value{
A matrix with \code{n} rows and \code{n} columns.
}
\description{
Creates the cross-covariance matrix of an AR(1) process with parameters
\code{rho} and \code{sigma}, observed at (positive) integer times
\code{times1} and \code{times2}.  The process is assumed to be in
stationarity and have Gaussian errors.
}
