% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar1_lpdf_cpp}
\alias{ar1_lpdf_cpp}
\title{Evaluate the log-density of a stationary Gaussian AR(1) process.}
\usage{
ar1_lpdf_cpp(x, mu, times, rho, sigma)
}
\arguments{
\item{x}{A vector of observed values.}

\item{mu}{A vector of expected values.}

\item{times}{A vector of the time points of observation.}

\item{rho}{A real number strictly less than 1 in absolute value.}

\item{sigma}{A positive real number.}
}
\value{
A scalar, the log density.
}
\description{
Evaluate the log-density of a stationary Gaussian AR(1) process, observed at
times \code{times} taking values \code{x}.
}
\keyword{internal}
