% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{ar1_sim}
\alias{ar1_sim}
\title{Simulate from a stationary Gaussian AR(1) process.}
\usage{
ar1_sim(times, rho, sigma, mu = 0)
}
\arguments{
\item{times}{The time points to simulate for. If a single positive integer,
then that many consecutive values will be simulated.}

\item{rho}{A real number strictly less than 1 in absolute value.}

\item{sigma}{A positive real number.}

\item{mu}{A vector of expected values with length \code{length(times)}, or a
scalar (default equal to 0).}
}
\value{
If \code{length(times) > 1}, a vector of length \code{length(times)}.
  Else (\code{length(times) == 1} a vector with as many values as the integer
  \code{times}.
}
\description{
Simulate from a stationary Gaussian AR(1) process, either at consecutive or
irregular times.
}
\examples{
times <- c(3, 5:7, 10)
rho <- 0.5
sigma <- 1
mu <- seq_along(times)
ar1_sim(times[1], rho, sigma)
ar1_sim(times, rho, sigma)
ar1_sim(times, rho, sigma, mu)
}
