% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar1_sim_cpp}
\alias{ar1_sim_cpp}
\title{Simulate from a stationary Gaussian AR(1) process.}
\usage{
ar1_sim_cpp(n, rho, sigma)
}
\arguments{
\item{n}{The number of timepoints to simulate for.}

\item{rho}{A real number strictly less than 1 in absolute value.}

\item{sigma}{A positive real number.}
}
\value{
A vector of length \code{n} with the process values.
}
\description{
Simulate from a stationary Gaussian AR(1) process at \code{n} consecutive
time points.
}
\keyword{internal}
