% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{band1_backsolve_mat}
\alias{band1_backsolve_mat}
\title{Backward substitution with band 1 lower Cholesky triangle and tridiagonal
RHS.}
\usage{
band1_backsolve_mat(L, Q)
}
\arguments{
\item{L}{A lower triangular square matrix with non-zero entries only on the
main diagonal and the first subdiagonal.}

\item{Q}{A tridiagonal matrix with the same dimensions as L.}
}
\value{
A matrix.
}
\description{
Backward substitution with band 1 lower Cholesky triangle and tridiagonal
matrix on the right hand side.
}
\keyword{internal}
