% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dcovchol_drho}
\alias{dcovchol_drho}
\title{Derivative of the inverse upper Cholesky triangle of the precision matrix of
a stationary Gaussian AR(1) process.}
\usage{
dcovchol_drho(U, dQ)
}
\arguments{
\item{U}{The upper Cholesky triangle of the precision matrix \code{Q} of the
AR(1) process.}

\item{dQ}{The derivative of the precision matrix \code{Q} with respect to
the correlation parameter \code{rho}.}
}
\value{
A matrix of the same dimensions as \code{U}.
}
\description{
Derivative of the inverse upper Cholesky triangle of the precision matrix of
of an AR(1) process with respect to the parameter \code{rho}.
}
\examples{
t <- c(1, 3:4, 6, 8)
r <- 0.5
s <- 1
U <- ar1_prec_chol_irregular(t, r, s)
dQ <- dprec_drho(t, r, s)
(dU <- dcovchol_drho(U, dQ))
}
\keyword{internal}
