% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mult_U_band1U}
\alias{mult_U_band1U}
\title{Multiply an upper triangular matrix with a band 1 upper triangular matrix.}
\usage{
mult_U_band1U(A, U)
}
\arguments{
\item{A}{A sparse upper triangular matrix.}

\item{U}{A sparse band 1 upper triangular matrix of the same dimensions as
\code{A}.}
}
\value{
A sparse band 1 upper triangular matrix.
}
\description{
Multiply an upper triangular matrix with a band 1 upper triangular matrix.
}
