% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trid_chol}
\alias{trid_chol}
\title{Cholesky decomposition of tridiagonal matrix
If Q is a tridiagonal matrix and Q = LL' = U'U, this function returns L'=U.
See Algorithm 2.9 of Rue & Held}
\usage{
trid_chol(Q)
}
\description{
Cholesky decomposition of tridiagonal matrix
If Q is a tridiagonal matrix and Q = LL' = U'U, this function returns L'=U.
See Algorithm 2.9 of Rue & Held
}
