% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar1_cov_chol_irregular}
\alias{ar1_cov_chol_irregular}
\title{Upper triangular Cholesky decomposition for a stationary Gaussian AR(1)
process, observed at irregularly spaced time points.}
\usage{
ar1_cov_chol_irregular(times, rho, sigma)
}
\arguments{
\item{times}{An vector of positive integers, preferably ordered.}

\item{rho}{A real number strictly less than 1 in absolute value.}

\item{sigma}{A positive real number.}
}
\value{
A square matrix with \code{length(times)} rows.
}
\description{
Creates the upper triangular Cholesky decomposition matrix of an AR(1)
process with parameters \code{rho} and \code{sigma}, observed at the time
points in the vector \code{times}. The process is assumed to be in
stationarity and have Gaussian errors.
}
