% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar1_prec_consecutive}
\alias{ar1_prec_consecutive}
\title{Precision matrix of vector x observed at n consecutive times.}
\usage{
ar1_prec_consecutive(n, rho, sigma)
}
\description{
Precision matrix of vector x observed at n consecutive times.
}
