% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar1_sim_irregular_cpp}
\alias{ar1_sim_irregular_cpp}
\title{Simulate from a stationary Gaussian AR(1) process at irregular times.}
\usage{
ar1_sim_irregular_cpp(times, rho, sigma)
}
\arguments{
\item{rho}{A real number strictly less than 1 in absolute value.}

\item{sigma}{A positive real number.}

\item{n}{The number of timepoints to simulate for.}
}
\value{
A vector of length \code{n} with the process values.
}
\description{
Simulate from a stationary Gaussian AR(1) process at irregular times.
}
\keyword{internal}
