% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{band1_backsolve}
\alias{band1_backsolve}
\title{Backsolve with band 1 upper Cholesky.}
\usage{
band1_backsolve(U, z)
}
\arguments{
\item{Q}{A square tridiagonal matrix.}
}
\value{
A sparse square matrix with the same size as the input matrix.
}
\description{
Backsolve with band 1 upper Cholesky.
}
